FIX: Verify correct PHP version before loading the plugin
